adds awaiting for worker task on cancellation
