+++ src/PSNetScanners/Abstractions/WorkerBase.cs
