+++ src/PSNetScanners/Cancellation.cs
@@ -6,17 +6,13 @@ namespace PSNetScanners;
 
 internal sealed class Cancellation : IDisposable
 {
-    private readonly CancellationTokenSource _cts;
+    private readonly CancellationTokenSource _cts = new();
 
     internal CancellationToken Token { get => _cts.Token; }
 
     internal Task Task { get; }
 
-    internal Cancellation()
-    {
-        _cts = new CancellationTokenSource();
-        Task = Task.Delay(Timeout.Infinite, _cts.Token);
-    }
+    internal Cancellation() => Task = Task.Delay(Timeout.Infinite, _cts.Token);
 
     internal Task GetTimeoutTask(int timeout) => Task.Delay(timeout, Token);
