+++ src/PSNetScanners/IWorker.cs
@@ -7,8 +7,8 @@ internal interface IWorker<TInput> : IDisposable
 {
     void Enqueue(TInput input);
     bool TryTake(out object result);
-    void Cancel();
     void CompleteAdding();
-    void Wait();
     IEnumerable<object> EnumerateOutput();
+    void Wait();
+    void CancelAndWait();
 }
